hotel-infer/requirements.txt
Flask==2.0.3
torch==1.13.1
torchvision==0.14.1
Pillow==9.0.1
numpy==1.22.4
flask-uploads==0.2.1
gunicorn==20.1.0
debugpy==1.6.3
matplotlib==3.5.3